requirements.txt
gradio
groq
python-dotenv
SpeechRecognition
gtts
pydub

